Fix bug 17302 - _close_statement is unable to free active statements
Overview
--------

During the termination of a result resource the driver
incorrectly nullifies the statement handle. This occurs
when removing the result resource from the associated
connection resource if the result statement handle
matches the link statement handle. This change makes it
so that php_ii_result_remove() does not nullify
ii_result->stmtHandle, leaving _close_statement() to
properly free/release the active statement and set
ii_result->stmtHandle to NULL.