Bug 17510 - Fix php_ii_set_connect_options so they work as documented
Overview
--------

The PHP driver provides the ability for users to set a number of connection
options when initiating a connection to a database. These options allow an
application to override the default environment settings inherited from the
environment. OpenAPI provides this capability with the IIapi_setConnectParam()
and IIapi_setEnvParam() functions, however the driver was not using these
functions correctly to set the relevant options. Whilst reviewing the test
suite for the driver it was discovered that the driver was not working as
expected.

This change updates php_ii_set_connect_options() so that it sets the relevant
option passed by ingres_connect()/ingres_pconnect(). Previously only certain
options would be processed by php_ii_set_connect_options(). Options not
handled by this function would be passed on to
php_ii_set_environment_options().  In addition this change fixes the driver
so that parameters whose value is a string, e.g. money_sign, now work.
The string values from PHP were not null terminated as a consequence were
being ignored. Now the driver allocates memory for each string parameter
value allowing for the null character. After each call to
IIapi_setConnectParam() this memory is released.

This change also simplifies the date_format and timezone tests. Until now
there were a number of tests which were exactly the same, save for a single
value, testing different timezones. Now these tests use a single test program
with an environment variable that defines the timezone or date_format used,
reducing future test maintenance.