Update the unit tests to be more independent
Overview
--------

Over time the unit tests added to the PHP driver have relied on
databases and/or tables which have existed in the core development
system. This change updates a number of tests so that they can be
run in a new Ingres installation with minimal changes. Anyone
wishing to run the tests will still need to set the ingres password
in tests/connection.inc.

It should be noted that with this change all the tests should complete
with the exception of the following:

tests/test_017.phpt
tests/test_048.phpt
tests/test_049.phpt

test_017.phpt tests the ability to connect with a dbms_password. Whilst
it's possible to temporarily set the dbms password on the ingres account
it's not something I'd like to do without having a way off un-setting the
password without relying on the ingres account.

tests/test_048.phpt and tests/test_048.phpt are failing due to memory
corruption. Valgrind is reporting a memory overlay when performing memcpy
during IIapi_wait(). This will be delt with at a later date.