Fix bug 17621 - Improve the handling of C style comments within an SQL query
Overview
--------

The PHP driver for Ingres is unable to deal with a parameter marker
that is embedded within:
- single quotes
- double quotes
- C style comments, /* */

As such the functions php_ii_paramcount() and php_ii_convert_query()
consider every ? to be a parameter marker irrespective of the context.
This change improves the parsing of the user query such that only
embedded markers are considered for parameter binding.

This change adds a new test, tests/test_056.phpt, to test the different
places a ? marker could be found. The query used is based on the SQL
found in http://bit.ly/IngresDriverComments