Fixed the scoring of results in search auto-complete. This required an upgrade to jQuery 1.4.4 and the inclusion of jQuery UI 1.8.7 (both linked to Google's Ajax APIs CDN) as the auto-complete plugin has been rolled into the standard jQuery UI library.
We now supply a source callback function to the auto-complete plugin (instead
of a JS array). This callback mimics the quick-ref logic for ranking manual 
pages based on where the user's search term appears in the name of the doc 
entry. However, it lacks the similar_text() component since that function is 
not available in JS (doesn't seem like a significant loss).

Tweaked auto-complete related CSS rules to accommodate different markup 
produced by the new version of the auto-complete plugin.