Avoid sending empty lines.
This change modifies the previous patch to explicitly check the length of the
string that was read from the stream.  The last fread() operation can return
an empty string, and passing an empty string to _send() can result in a socket
error.

We need to use strlen() here instead of just testing ($line) because PHP
evaluates strings like "0" as FALSE.