Fixed bug #63491 file_md5 value was wrong when use apc_bin_load function
The basic idea behind this is - original file md5 isn't dumped, as result
loading that dump the key->md5 is filled with the garbage (which could 
look like a wrong md5 sum) as we don't read the file again with 
cached_compile. What i did:

- zero the original file md5 on key creation
- write the file md5 into the bin dump when apc_bin_dumpfile()'ing
- restore the md5 field when loading the dump with apc_dump_loadfile()

However we only calculate the md5 sum if apc.file_md5=1, so otherwise
files with the preloaded dump will have just zeros.