Improve utf8_decode() and utf8_encode() documentation
This rewrites the descriptions of both to clarify that they convert 
specifically between ISO-8859-1 and UTF-8, adds a warning about
confusion with Windows-1252, and adds helpful "See also" links to
other character set conversion functions. Additionally, the
behaviour for invalid characters in utf8_decode() was clarified,
and the description of the UTF-8 binary encoding was removed.