Use proper resources instead of integer identifiers
The database handles should be really opaque to userland, and not only
treated as such, to avoid confusion and bugs. This appears to be
particularly relevant as of PHP 7.0.0 which introduces `int` type hints.

We consider the BC break to be relatively small; quite likely only some
tests would be affected (as can be seen in our test suite). And, anyhow,
we're working on a new major version, so at least some BC breaks are to
be expected.

Objects would be an alternative, but without methods they won't be a
real improvement over resources, so we leave that for a future version.