Add support for datetime fields
We add support for FoxPro datetime fields, which store an exact
timestamp in two binary dwords, representing the Julian day number and
the number of milliseconds since midnight, respectively. These are
represented in userland as strings with the format "YYYYMMDDhhmmss.uuu",
which conforms to ISO-8601.

According to the current handling of date fields, the validity of input
and output is not checked, i.e. it is possible to store invalid datetime
values in the database. If necessary, respective checks have to be made
by the programmer.

To avoid reinventing the wheel, but also not to rely on ext/calendar, we
add the respective functionality and adapt the names to avoid build
conflicts.