Implement basic database locking
We're using advisory file locking via php_flock(), so the usual caveats
regarding multithreaded Windows SAPIs apply. We're applying blocking
locks automatically on creation and opening of databases, so that
existing code doesn't have to be changed to make use of the locking. If
users want to check whether an operation would block, flock() with
LOCK_NB can be used (leaving room for possible race-conditions, though).