Add transparent support for nullable fields
Nullable fields are a FoxPro extension to DBF, and are implemented by
adding a _NullFlags field of type '0' which stores a bitmask indicating
which fields of the record are NULL.

We're adding transparent support for nullable fields, i.e. the
_NullFlags field is never shown to user, but rather when storing records
with NULL values the _NullFlags are set appropriately, and when
retrieving records with NULL values the respective elements are set to
NULL.

When creating a new DBF with dbase_create() the field specifications
accept a further optional element (default is FALSE) that indicates
whether the field is nullable. These nullable elements are ignored for
DBASE_TYPE_DBASE.